+++ src/shillelagh/adapters/api/generic_json.py
@@ -22,17 +22,21 @@ _logger = logging.getLogger(__name__)
 
 SUPPORTED_PROTOCOLS = {"http", "https"}
 AVERAGE_NUMBER_OF_ROWS = 100
+CACHE_EXPIRATION = 180
 
 
-def get_session() -> requests_cache.CachedSession:
+def get_session(request_headers: Dict[str, str]) -> requests_cache.CachedSession:
     """
     Return a cached session.
     """
-    return requests_cache.CachedSession(
+    session = requests_cache.CachedSession(
         cache_name="generic_json_cache",
         backend="sqlite",
-        expire_after=180,
+        expire_after=CACHE_EXPIRATION,
     )
+    session.headers.update(request_headers)
+
+    return session
 
 
 class GenericJSONAPI(Adapter):
@@ -54,7 +58,8 @@ class GenericJSONAPI(Adapter):
         if fast:
             return Maybe
 
-        session = get_session()
+        request_headers = kwargs.get("request_headers", {})
+        session = get_session(request_headers)
         response = session.head(uri)
         return "application/json" in response.headers.get("content-type", "")
 
@@ -67,13 +72,18 @@ class GenericJSONAPI(Adapter):
 
         return uri, path
 
-    def __init__(self, uri: str, path: str = "$[*]"):
+    def __init__(
+        self,
+        uri: str,
+        path: str = "$[*]",
+        request_headers: Optional[Dict[str, str]] = None,
+    ):
         super().__init__()
 
         self.uri = uri
         self.path = path
 
-        self._session = get_session()
+        self._session = get_session(request_headers or {})
 
         self._set_columns()